requirements.txt
django==5.0.3
django-environ==0.11.2
psycopg2-binary==2.9.9
django-money==3.4.1
django-livereload-server
pymongo==4.6.2
mongoengine==0.28.2
snowflake-connector-python==3.7.1
redis==5.0.3